Fixed bug with global ordinal of long

+++ packages/muelu/src/Rebalancing/MueLu_Zoltan2Interface_def.hpp
@@ -164,7 +164,7 @@ namespace MueLu {
 
     } else if (algo == "rcb") {
       params.set("algorithm",             "rcb");
-      params.set("num_global_parts",      numParts);
+      params.set("num_global_parts",      static_cast<int>(numParts));
     }
 
     params.set("partitioning_approach",   "partition");
